Added the games screen
