+++ src/modules/games/components/card.js
@@ -1,12 +1,32 @@
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
+import { LinkContainer } from 'react-router-bootstrap'
+import { Col, Row, Panel, Image } from 'react-bootstrap'
 
 class GameCard extends PureComponent {
 
   render() {
     const { game } = this.props
 
-    return <div>{this.props.game.name}</div>
+    return (
+      <Panel bsStyle="primary">
+        <Panel.Heading>{game.name}</Panel.Heading>
+          <Panel.Body>
+            <Row>
+              <Col xs={12} md={7}>
+                <LinkContainer to={`/games/${game.id}`}>
+                  <Image className='game-avatar' src={game.img_url} />
+                </LinkContainer>
+              </Col>
+              <Col xs={12} md={5}>
+                <div>Players: {game.players}</div>
+                <div>Scoring: TODO</div>
+                <div>Description: TODO</div>
+              </Col>
+            </Row>
+          </Panel.Body>
+      </Panel>
+    )
   }
 }
