+++ src/modules/games/components/list.js
@@ -3,8 +3,24 @@ import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { values } from 'lodash/object'
+import { chunk } from 'lodash/array'
+import { Grid, Col, Row } from 'react-bootstrap'
 
+import GameCard from '../components/card'
 import { fetchGames } from '../actions'
+
+const GameRow = ({ row }) => (
+  row.map((game) => (
+    <Col key={game.id} md={6} xs={12}>
+      <GameCard style={{ height: '100%' }} game={game}/>
+    </Col>
+  ))
+)
+
+const GameCards = ({ games, columns = 2 }) => {
+  const rows = chunk(games, 2)
+  return rows.map(row => <Row><GameRow row={row}/></Row>)
+}
 
 class GameDetail extends PureComponent {
 
@@ -12,32 +28,15 @@ class GameDetail extends PureComponent {
     this.props.actions.fetchGames()
   }
 
-  renderData() {
-    const { games } = this.props
-
-    return (
-      <ul>
-        {
-          games.map((t) =>
-            <li key={t.id}>{t.name}</li>
-          )
-        }
-      </ul>
-    )
-  }
-
   render() {
     const { games } = this.props
 
+    if (games.length <= 0) {
+      return null
+    }
+
     return (
-      <div>
-        { games.length > 0 ?
-          this.renderData() :
-          <div className="">
-            No Data
-          </div>
-        }
-      </div>
+      <Grid><GameCards games={games} /></Grid>
     )
   }
 }
