+++ src/modules/players/views/index.js
@@ -1,3 +1,3 @@
-import PlayerList from '../components/detail'
+import PlayerList from '../components/list'
 
 export default PlayerList
